backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# API Clients
httpx==0.27.2
anthropic==0.39.0

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
